% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_example.R
\name{create_example}
\alias{create_example}
\title{Create Example}
\usage{
create_example(
  progression_table = RIR_increment,
  reps = c(3, 5, 10),
  volume = c("intensive", "normal", "extensive"),
  ...
)
}
\arguments{
\item{progression_table}{Progression table function. Default is \code{\link{RIR_increment}}}

\item{reps}{Numeric vector. Default is \code{c(3, 5, 10)}}

\item{volume}{Character vector. Default is \code{c("intensive", "normal", "extensive")}}

\item{...}{Extra arguments forwarded to \code{progression_table}}
}
\description{
This function create simple example using \code{progression_table}
}
\examples{
create_example(perc_drop)

# Create example using specific reps-max table and k value
create_example(
  RIR_increment,
  func_max_perc_1RM = get_max_perc_1RM_k,
  k = 0.0388)

}
