% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-reps-table-k.R, R/max-reps-table-klin.R,
%   R/max-reps-table-kmod.R, R/max-reps-table-percMR.R,
%   R/max-reps-table-relInt.R, R/max-reps-table.R
\name{get_max_reps_k}
\alias{get_max_reps_k}
\alias{get_max_perc_1RM_k}
\alias{get_predicted_1RM_k}
\alias{get_max_perc_1RM_k_relInt}
\alias{get_max_perc_1RM_k_percMR}
\alias{get_max_reps_klin}
\alias{get_max_perc_1RM_klin}
\alias{get_predicted_1RM_klin}
\alias{get_max_perc_1RM_klin_relInt}
\alias{get_max_perc_1RM_klin_percMR}
\alias{get_max_reps_kmod}
\alias{get_max_perc_1RM_kmod}
\alias{get_predicted_1RM_kmod}
\alias{get_max_perc_1RM_kmod_relInt}
\alias{get_max_perc_1RM_kmod_percMR}
\alias{get_max_reps_percMR}
\alias{get_max_perc_1RM_percMR}
\alias{get_predicted_1RM_percMR}
\alias{get_max_reps_relInt}
\alias{get_max_perc_1RM_relInt}
\alias{max_reps_tables}
\alias{get_max_reps}
\alias{get_max_perc_1RM}
\alias{get_predicted_1RM}
\title{Family of functions to represent max reps tables}
\usage{
get_max_reps_k(perc_1RM, adjustment = 0, type = "grinding", k = 0.0333)

get_max_perc_1RM_k(max_reps, adjustment = 0, type = "grinding", k = 0.0333)

get_predicted_1RM_k(
  weight,
  reps,
  adjustment = 0,
  type = "grinding",
  k = 0.0333
)

get_max_perc_1RM_k_relInt(
  max_reps,
  adjustment = 1,
  type = "grinding",
  k = 0.0333
)

get_max_perc_1RM_k_percMR(
  max_reps,
  adjustment = 1,
  type = "grinding",
  k = 0.0333
)

get_max_reps_klin(perc_1RM, adjustment = 0, type = "grinding", klin = 33)

get_max_perc_1RM_klin(max_reps, adjustment = 0, type = "grinding", klin = 33)

get_predicted_1RM_klin(
  weight,
  reps,
  adjustment = 0,
  type = "grinding",
  klin = 33
)

get_max_perc_1RM_klin_relInt(
  max_reps,
  adjustment = 1,
  type = "grinding",
  klin = 33
)

get_max_perc_1RM_klin_percMR(
  max_reps,
  adjustment = 1,
  type = "grinding",
  klin = 33
)

get_max_reps_kmod(perc_1RM, adjustment = 0, type = "grinding", kmod = 0.0353)

get_max_perc_1RM_kmod(
  max_reps,
  adjustment = 0,
  type = "grinding",
  kmod = 0.0353
)

get_predicted_1RM_kmod(
  weight,
  reps,
  adjustment = 0,
  type = "grinding",
  kmod = 0.0353
)

get_max_perc_1RM_kmod_relInt(
  max_reps,
  adjustment = 1,
  type = "grinding",
  kmod = 0.0353
)

get_max_perc_1RM_kmod_percMR(
  max_reps,
  adjustment = 1,
  type = "grinding",
  kmod = 0.0353
)

get_max_reps_percMR(perc_1RM, adjustment = 1, type = "grinding")

get_max_perc_1RM_percMR(max_reps, adjustment = 1, type = "grinding")

get_predicted_1RM_percMR(weight, reps, adjustment = 1, type = "grinding")

get_max_reps_relInt(perc_1RM, adjustment = 1, type = "grinding")

get_max_perc_1RM_relInt(max_reps, adjustment = 1, type = "grinding")

get_max_reps(perc_1RM, adjustment = 0, type = "grinding")

get_max_perc_1RM(max_reps, adjustment = 0, type = "grinding")

get_predicted_1RM(weight, reps, adjustment = 0, type = "grinding")
}
\arguments{
\item{perc_1RM}{Percent of 1RM}

\item{adjustment}{Adjustment to be implemented, specific to the method}

\item{type}{Type of max rep table. Options are grinding (Default) and ballistic.}

\item{k}{User defined \code{k} parameter in the Epley's equation. Default is 0.0333}

\item{max_reps}{Maximum repetitions}

\item{weight}{Weight used}

\item{reps}{Number of repetitions done}

\item{klin}{User defined \code{klin} parameter in the linear equation. Default is 33}

\item{kmod}{User defined \code{kmod} parameter in the modified Epley's equation. Default is 0.0353}
}
\description{
Family of functions to represent max reps tables
}
\section{Functions}{
\itemize{
\item \code{get_max_reps_k}: Get maximum number of repetitions using the RIR method (\code{adjustment}), but
with user provided \code{k} parameter

\item \code{get_max_perc_1RM_k}: Get maximum \%1RM using the RIR method (\code{adjustment}), but
with user provided \code{k} parameter

\item \code{get_predicted_1RM_k}: Get predicted 1RM using the RIR method (\code{adjustment}), but
with user provided \code{k} parameter

\item \code{get_max_perc_1RM_k_relInt}: Get maximum \%1RM using the Relative Intensity (\code{adjustment}), but
with user provided \code{k} parameter

\item \code{get_max_perc_1RM_k_percMR}: Get maximum \%1RM using the \%MR method (\code{adjustment}), but
with user provided \code{k} parameter

\item \code{get_max_reps_klin}: Get maximum number of repetitions using the RIR method (\code{adjustment}), but
with user provided \code{klin} parameter using linear equation

\item \code{get_max_perc_1RM_klin}: Get maximum \%1RM using the RIR method (\code{adjustment}), but
with user provided \code{klin} parameter using linear equation

\item \code{get_predicted_1RM_klin}: Get predicted 1RM using the RIR method (\code{adjustment}), but
with user provided \code{klin} parameter using linear equation

\item \code{get_max_perc_1RM_klin_relInt}: Get maximum \%1RM using the Relative Intensity (\code{adjustment}), but
with user provided \code{klin} parameter using linear equation

\item \code{get_max_perc_1RM_klin_percMR}: Get maximum \%1RM using the \%MR (\code{adjustment}), but
with user provided \code{klin} parameter using linear equation

\item \code{get_max_reps_kmod}: Get maximum number of repetitions using the RIR method (\code{adjustment}), but
with user provided \code{kmod} parameter using modified Epley's equation

\item \code{get_max_perc_1RM_kmod}: Get maximum \%1RM using the RIR method (\code{adjustment}), but
with user provided \code{kmod} parameter using modified Epley's equation

\item \code{get_predicted_1RM_kmod}: Get predicted 1RM using the RIR method (\code{adjustment}), but
with user provided \code{kmod} parameter using modified Epley's equation

\item \code{get_max_perc_1RM_kmod_relInt}: Get maximum \%1RM using the Relative Intensity (\code{adjustment}), but
with user provided \code{kmod} parameter

\item \code{get_max_perc_1RM_kmod_percMR}: Get maximum \%1RM using the \%MR method (\code{adjustment}), but
with user provided \code{kmod} parameter

\item \code{get_max_reps_percMR}: Get maximum number of repetitions using the \%MR (\code{adjustment})

\item \code{get_max_perc_1RM_percMR}: Get maximum \%1RM using the \%MR (\code{adjustment})

\item \code{get_predicted_1RM_percMR}: Get predicted 1RM using the \%MR (\code{adjustment})

\item \code{get_max_reps_relInt}: Get maximum number of repetitions using the Relative Intensity (\code{adjustment})

\item \code{get_max_perc_1RM_relInt}: Get maximum \%1RM using the Relative Intensity (\code{adjustment})

\item \code{get_max_reps}: Get maximum number of repetitions using the RIR method (\code{adjustment})

\item \code{get_max_perc_1RM}: Get maximum \%1RM using the RIR method (\code{adjustment})

\item \code{get_predicted_1RM}: Get predicted 1RM using the RIR method (\code{adjustment})
}}

\examples{
# Get max reps that can be done with 75\% 1RM
get_max_reps_k(0.75, k = 0.05)

# Get max reps that can be done with 80\% 1RM with 2 reps in reserve
get_max_reps_k(
  perc_1RM = 0.8,
  adjustment = 3,
  k = 0.05
)

# Get max reps that can be done with 90\% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps_k(
  perc_1RM = 0.9,
  adjustment = 2,
  type = "ballistic",
  k = 0.05
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_k(5, k = 0.0333)

# Get max \%1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM_k(
  max_reps = 3,
  adjustment = 2,
  k = 0.05
)

# Get max \%1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM_k(
  max_reps = 3,
  adjustment = 2,
  type = "ballistic",
  k = 0.05
)
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM_k(
  weight = 100,
  reps = 5,
  k = 0.05
)

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM_k(
  weight = 120,
  reps = 3,
  adjustment = 2,
  k = 0.05
)

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM_k(
  weight = 120,
  reps = 2,
  adjustment = 1,
  type = "ballistic",
  k = 0.05
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_k_relInt(5, k = 0.0333)

# Get max \%1RM to be used when doing 3 reps with 60\% of Relative Intensity
get_max_perc_1RM_k_relInt(
  max_reps = 3,
  adjustment = 0.6,
  k = 0.05
)

# Get max \%1RM to be used when doing 2 reps with 60\% of Relative Intensity
# using ballistic table
get_max_perc_1RM_k_relInt(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic",
  k = 0.05
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_k_percMR(5, k = 0.0333)

# Get max \%1RM to be used when doing 3 reps with 60\% of maximum reps
get_max_perc_1RM_k_percMR(
  max_reps = 3,
  adjustment = 0.6,
  k = 0.05
)

# Get max \%1RM to be used when doing 2 reps with 60\% of maximum reps
# using ballistic table
get_max_perc_1RM_k_percMR(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic",
  k = 0.05
)
# Get max reps that can be done with 75\% 1RM
get_max_reps_klin(0.75, klin = 33)

# Get max reps that can be done with 80\% 1RM with 2 reps in reserve
get_max_reps_klin(
  perc_1RM = 0.8,
  adjustment = 3,
  klin = 33
)

# Get max reps that can be done with 90\% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps_klin(
  perc_1RM = 0.9,
  adjustment = 2,
  type = "ballistic",
  klin = 33
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_klin(5, klin = 33)

# Get max \%1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM_klin(
  max_reps = 3,
  adjustment = 2,
  klin = 33
)

# Get max \%1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM_klin(
  max_reps = 3,
  adjustment = 2,
  type = "ballistic",
  klin = 33
)
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM_klin(
  weight = 100,
  reps = 5,
  klin = 33
)

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM_klin(
  weight = 120,
  reps = 3,
  adjustment = 2,
  klin = 33
)

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM_klin(
  weight = 120,
  reps = 2,
  adjustment = 1,
  type = "ballistic",
  klin = 33
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_klin_relInt(5, klin = 35)

# Get max \%1RM to be used when doing 3 reps with 60\% of Relative Intensity
get_max_perc_1RM_klin_relInt(
  max_reps = 3,
  adjustment = 0.6,
  klin = 35
)

# Get max \%1RM to be used when doing 2 reps with 60\% of Relative Intensity
# using ballistic table
get_max_perc_1RM_klin_relInt(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic",
  klin = 35
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_klin_percMR(5, klin = 35)

# Get max \%1RM to be used when doing 3 reps with 60\% of maximum reps
get_max_perc_1RM_klin_percMR(
  max_reps = 3,
  adjustment = 0.6,
  klin = 35
)

# Get max \%1RM to be used when doing 2 reps with 60\% of maximum reps
# using ballistic table
get_max_perc_1RM_klin_percMR(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic",
  klin = 35
)
# Get max reps that can be done with 75\% 1RM
get_max_reps_kmod(0.75, kmod = 0.05)

# Get max reps that can be done with 80\% 1RM with 2 reps in reserve
get_max_reps_kmod(
  perc_1RM = 0.8,
  adjustment = 3,
  kmod = 0.05
)

# Get max reps that can be done with 90\% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps_kmod(
  perc_1RM = 0.9,
  adjustment = 2,
  type = "ballistic",
  kmod = 0.05
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_kmod(5, kmod = 0.0353)

# Get max \%1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM_kmod(
  max_reps = 3,
  adjustment = 2,
  kmod = 0.05
)

# Get max \%1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM_kmod(
  max_reps = 3,
  adjustment = 2,
  type = "ballistic",
  kmod = 0.05
)
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM_kmod(
  weight = 100,
  reps = 5,
  kmod = 0.05
)

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM_kmod(
  weight = 120,
  reps = 3,
  adjustment = 2,
  kmod = 0.05
)

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM_kmod(
  weight = 120,
  reps = 2,
  adjustment = 1,
  type = "ballistic",
  kmod = 0.05
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_kmod_relInt(5, kmod = 0.0333)

# Get max \%1RM to be used when doing 3 reps with 60\% of Relative Intensity
get_max_perc_1RM_kmod_relInt(
  max_reps = 3,
  adjustment = 0.6,
  kmod = 0.05
)

# Get max \%1RM to be used when doing 2 reps with 60\% of Relative Intensity
# using ballistic table
get_max_perc_1RM_kmod_relInt(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic",
  kmod = 0.05
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_kmod_percMR(5, kmod = 0.0353)

# Get max \%1RM to be used when doing 3 reps with 60\% of maximum reps
get_max_perc_1RM_kmod_percMR(
  max_reps = 3,
  adjustment = 0.6,
  kmod = 0.05
)

# Get max \%1RM to be used when doing 2 reps with 60\% of maximum reps
# using ballistic table
get_max_perc_1RM_kmod_percMR(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic",
  kmod = 0.05
)
# Get max reps that can be done with 75\% 1RM
get_max_reps_percMR(0.75)

# Get max reps that can be done with 80\% 1RM with 60\% of maximum reps
get_max_reps_percMR(
  perc_1RM = 0.8,
  adjustment = 0.6
)

# Get max reps that can be done with 90\% 1RM with 60\% of maximum reps
# using ballistic table
get_max_reps_percMR(
  perc_1RM = 0.9,
  adjustment = 0.6,
  type = "ballistic"
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_percMR(5)

# Get max \%1RM to be used when doing 3 reps with with 60\% of maximum reps
get_max_perc_1RM_percMR(
  max_reps = 3,
  adjustment = 0.6,
)

# Get max \%1RM to be used when doing 2 reps with 60\% of maximum reps
# using ballistic table
get_max_perc_1RM_percMR(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic"
)
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM_percMR(
  weight = 100,
  reps = 5
)

# Get predicted 1RM when lifting 120kg for 3 reps with 60\% of maximum reps
get_predicted_1RM_percMR(
  weight = 120,
  reps = 3,
  adjustment = 0.6
)

# Get predicted 1RM when lifting 120kg for 2 reps with 60\% of maximum reps
# using ballistic table
get_predicted_1RM_percMR(
  weight = 120,
  reps = 2,
  adjustment = 0.6,
  type = "ballistic"
)
# Get max reps that can be done with 75\% 1RM
get_max_reps_relInt(0.75)

# Get max reps that can be done with 80\% 1RM with 60\% of Relative Intensity
get_max_reps_relInt(
  perc_1RM = 0.8,
  adjustment = 0.6
)

# Get max reps that can be done with 90\% 1RM with 60\% of Relative Intensity
# using ballistic table
get_max_reps_relInt(
  perc_1RM = 0.9,
  adjustment = 0.6,
  type = "ballistic"
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_relInt(5)

# Get max \%1RM to be used when doing 3 reps with with 60\% of Relative Intensity
get_max_perc_1RM_relInt(
  max_reps = 3,
  adjustment = 0.6,
)

# Get max \%1RM to be used when doing 2 reps with 60\% of Relative Intensity
# using ballistic table
get_max_perc_1RM_relInt(
  max_reps = 3,
  adjustment = 0.6,
  type = "ballistic"
)
# Get max reps that can be done with 75\% 1RM
get_max_reps(0.75)

# Get max reps that can be done with 80\% 1RM with 2 reps in reserve
get_max_reps(
  perc_1RM = 0.8,
  adjustment = 3
)

# Get max reps that can be done with 90\% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps(
  perc_1RM = 0.9,
  adjustment = 2,
  type = "ballistic"
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM(5)

# Get max \%1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM(
  max_reps = 3,
  adjustment = 2
)

# Get max \%1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM(
  max_reps = 3,
  adjustment = 2,
  type = "ballistic"
)
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM(
  weight = 100,
  reps = 5
)

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM(
  weight = 120,
  reps = 3,
  adjustment = 2
)

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM(
  weight = 120,
  reps = 2,
  adjustment = 1,
  type = "ballistic"
)
}
