% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vertical.R
\name{plot_vertical}
\alias{plot_vertical}
\title{Plotting of the Vertical Planning}
\usage{
plot_vertical(vertical_plan, reps = c(5, 5, 5), font_size = 14, ...)
}
\arguments{
\item{vertical_plan}{Vertical Plan function}

\item{reps}{Numeric vector}

\item{font_size}{Numeric. Default is 14}

\item{...}{Forwarded to \code{vertical_plan} function}
}
\description{
Function for creating \code{ggplot2} plot of the Vertical Planning function
}
\examples{
plot_vertical(vertical_block_undulating, reps = c(8, 6, 4))
}
