% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemes-S3.R
\name{+.STMr_scheme}
\alias{+.STMr_scheme}
\title{Method for adding set and rep schemes}
\usage{
\method{+}{STMr_scheme}(lhs, rhs)
}
\arguments{
\item{lhs}{\code{STMr_scheme} object}

\item{rhs}{\code{STMr_scheme} object}
}
\value{
\code{STMr_scheme} object
}
\description{
Method for adding set and rep schemes
}
\examples{
scheme1 <- scheme_wave()
warmup_scheme <- scheme_perc_1RM()
plot(warmup_scheme + scheme1)
}
