% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sig_pad}
\alias{sig_pad}
\title{Format to significant digits and pad to equal string width}
\usage{
sig_pad(x, sig = 3L, na = NA_character_)
}
\arguments{
\item{x}{Numeric vector.}

\item{sig}{Integer >= 1. Significant digits.}

\item{na}{Character to use for NA values.}
}
\value{
Character vector with equal nchar (non-NA values), left-padded with spaces.
}
\description{
Format to significant digits and pad to equal string width
}
