% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-example.R
\name{create_example}
\alias{create_example}
\title{Create Example}
\usage{
create_example(
  progression_table = progression_RIR_increment,
  reps = c(3, 5, 10),
  volume = c("intensive", "normal", "extensive"),
  type = c("grinding", "ballistic"),
  ...
)
}
\arguments{
\item{progression_table}{Progression table function. Default is \code{\link{progression_RIR_increment}}}

\item{reps}{Numeric vector. Default is \code{c(3, 5, 10)}}

\item{volume}{Character vector. Default is \code{c("intensive", "normal", "extensive")}}

\item{type}{Character vector. Type of max rep table. Options are grinding (Default) and ballistic}

\item{...}{Extra arguments forwarded to \code{progression_table}}
}
\description{
This function create simple example using \code{progression_table}
}
\examples{
create_example(progression_RIR)

# Create example using specific reps-max table and k value
create_example(
  progression_RIR,
  max_perc_1RM_func = max_perc_1RM_modified_epley,
  kmod = 0.0388
)
}
