% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-functions.R
\name{estimate_functions}
\alias{estimate_functions}
\alias{estimate_k}
\alias{estimate_kmod}
\title{Estimate relationship between reps and \%1RM used using Epley's and modified
Epley's equation}
\usage{
estimate_k(perc_1RM, reps, adjustment = 0, ...)

estimate_kmod(perc_1RM, reps, adjustment = 0, ...)
}
\arguments{
\item{perc_1RM}{Percent of 1RM}

\item{reps}{Number of repetitions done}

\item{adjustment}{Adjustment to be implemented, specific to the method}

\item{...}{Forwarded to \code{\link[stats]{nls}} functions}
}
\value{
\code{\link[stats]{nls}} object
}
\description{
Estimate relationship between reps and \%1RM used using Epley's and modified
Epley's equation
}
\section{Functions}{
\itemize{
\item \code{estimate_k}: Estimate the parameter k in the Epley's equation

\item \code{estimate_kmod}: Estimate the parameter kmod in the modified Epley's equation
}}

\examples{

m1 <- estimate_k(
  perc_1RM = c(0.7, 0.8, 0.9),
  reps = c(10, 5, 3)
)

coef(m1)

m1 <- estimate_kmod(
  perc_1RM = c(0.7, 0.8, 0.9),
  reps = c(10, 5, 3)
)

coef(m1)
}
