% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-functions.R
\name{estimate_functions}
\alias{estimate_functions}
\alias{estimate_k}
\alias{estimate_kmod}
\alias{estimate_k_1RM}
\alias{estimate_kmod_1RM}
\alias{estimate_klin}
\alias{estimate_klin_1RM}
\title{Estimate relationship between reps and \%1RM used using Epley's and modified
Epley's equation, as well as using the linear model.}
\usage{
estimate_k(perc_1RM, reps, adjustment = 0, reverse = FALSE, ...)

estimate_kmod(perc_1RM, reps, adjustment = 0, reverse = FALSE, ...)

estimate_k_1RM(weight, reps, adjustment = 0, reverse = FALSE, ...)

estimate_kmod_1RM(weight, reps, adjustment = 0, reverse = FALSE, ...)

estimate_klin(perc_1RM, reps, adjustment = 0, reverse = FALSE, ...)

estimate_klin_1RM(weight, reps, adjustment = 0, reverse = FALSE, ...)
}
\arguments{
\item{perc_1RM}{Percent of 1RM}

\item{reps}{Number of repetitions done}

\item{adjustment}{Subjective estimation of reps-in-reserve (eRIR)}

\item{reverse}{Logical, default is \code{FALSE}. Should reps be used as predictor instead as a target?}

\item{...}{Forwarded to \code{\link[stats]{nls}} functions}

\item{weight}{Weight used}
}
\value{
\code{\link[stats]{nls}} object
}
\description{
By default, target variable is the reps performed, while the predictors is the \code{perc_1RM} or
\code{weight}. To reverse this, use the \code{reverse = TRUE} argument
}
\section{Functions}{
\itemize{
\item \code{estimate_k}: Estimate the parameter \code{k} in the Epley's equation

\item \code{estimate_kmod}: Estimate the parameter \code{kmod} in the modified Epley's equation

\item \code{estimate_k_1RM}: Estimate the parameter \code{k} in the Epley's equation, as well as
\code{1RM}. This is a novel estimation function that uses the absolute weights

\item \code{estimate_kmod_1RM}: Estimate the parameter \code{kmod} in the modified Epley's equation, as well as
\code{1RM}. This is a novel estimation function that uses the absolute weights

\item \code{estimate_klin}: Estimate the parameter \code{klin} using the linear model

\item \code{estimate_klin_1RM}: Estimate the parameter \code{klin} in the linear equation, as well as
\code{1RM}. This is a novel estimation function that uses the absolute weights
}}

\examples{

m1 <- estimate_k(
  perc_1RM = c(0.7, 0.8, 0.9),
  reps = c(10, 5, 3)
)

coef(m1)

m1 <- estimate_kmod(
  perc_1RM = c(0.7, 0.8, 0.9),
  reps = c(10, 5, 3)
)

coef(m1)

m1 <- estimate_k_1RM(
  weight = c(70, 110, 140),
  reps = c(10, 5, 3)
)

coef(m1)

m1 <- estimate_kmod_1RM(
  weight = c(70, 110, 140),
  reps = c(10, 5, 3)
)

coef(m1)

m1 <- estimate_klin(
  perc_1RM = c(0.7, 0.8, 0.9),
  reps = c(10, 5, 3)
)

coef(m1)

m1 <- estimate_klin_1RM(
  weight = c(70, 110, 140),
  reps = c(10, 5, 3)
)

coef(m1)
}
