% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-reps-tables-percMR.R, R/max-reps-tables.R
\name{get_max_reps_percMR}
\alias{get_max_reps_percMR}
\alias{get_max_perc_1RM_percMR}
\alias{get_predicted_1RM_percMR}
\alias{max_reps_tables}
\alias{get_max_reps}
\alias{get_max_perc_1RM}
\alias{get_predicted_1RM}
\title{Max-Reps Tables}
\usage{
get_max_reps_percMR(perc_1RM, percMR = 1, type = "grinding")

get_max_perc_1RM_percMR(max_reps, percMR = 1, type = "grinding")

get_predicted_1RM_percMR(weight, reps, percMR = 1, type = "grinding")

get_max_reps(perc_1RM, RIR = 0, type = "grinding")

get_max_perc_1RM(max_reps, RIR = 0, type = "grinding")

get_predicted_1RM(weight, reps, RIR = 0, type = "grinding")
}
\arguments{
\item{perc_1RM}{Percent of 1RM}

\item{percMR}{Percent of maximum repetitions}

\item{type}{Type of max rep table. Options are grinding (Default) and ballistic.}

\item{max_reps}{Maximum repetitions}

\item{weight}{Weight used}

\item{reps}{Number of repetitions done}

\item{RIR}{Reps In Reserve}
}
\description{
Family of functions to represent max reps tables
}
\section{Functions}{
\itemize{
\item \code{get_max_reps_percMR}: Get maximum number of repetitions using the \%MR (\code{percMR})

\item \code{get_max_perc_1RM_percMR}: Get maximum \%1RM using the \%MR (\code{percMR})

\item \code{get_predicted_1RM_percMR}: Get predicted 1RM using the \%MR (\code{percMR})

\item \code{get_max_reps}: Get maximum number of repetitions

\item \code{get_max_perc_1RM}: Get maximum \%1RM

\item \code{get_predicted_1RM}: Get predicted 1RM
}}

\examples{
# Get max reps that can be done with 75\% 1RM
get_max_reps_percMR(0.75)

# Get max reps that can be done with 80\% 1RM with 60\% of maximum reps
get_max_reps_percMR(
  perc_1RM = 0.8,
  percMR = 0.6
)

# Get max reps that can be done with 90\% 1RM with 60\% of maximum reps
# using ballistic table
get_max_reps_percMR(
  perc_1RM = 0.9,
  percMR = 0.6,
  type = "ballistic"
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM_percMR(5)

# Get max \%1RM to be used when doing 3 reps with with 60\% of maximum reps
get_max_perc_1RM_percMR(
  max_reps = 3,
  percMR = 0.6,
)

# Get max \%1RM to be used when doing 2 reps with 60\% of maximum reps
# using ballistic table
get_max_perc_1RM_percMR(
  max_reps = 3,
  percMR = 0.6,
  type = "ballistic"
)
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM_percMR(
  weight = 100,
  reps = 5
)

# Get predicted 1RM when lifting 120kg for 3 reps with 60\% of maximum reps
get_predicted_1RM_percMR(
  weight = 120,
  reps = 3,
  percMR = 0.6
)

# Get predicted 1RM when lifting 120kg for 2 reps with 60\% of maximum reps
# using ballistic table
get_predicted_1RM_percMR(
  weight = 120,
  reps = 2,
  percMR = 0.6,
  type = "ballistic"
)
# Get max reps that can be done with 75\% 1RM
get_max_reps(0.75)

# Get max reps that can be done with 80\% 1RM with 2 reps in reserve
get_max_reps(
  perc_1RM = 0.8,
  RIR = 3
)

# Get max reps that can be done with 90\% 1RM with 2 reps in reserve
# using ballistic table
get_max_reps(
  perc_1RM = 0.9,
  RIR = 2,
  type = "ballistic"
)
# Get max \%1RM to be used when doing 5 reps to failure
get_max_perc_1RM(5)

# Get max \%1RM to be used when doing 3 reps with 2 reps in reserve
get_max_perc_1RM(
  max_reps = 3,
  RIR = 2
)

# Get max \%1RM to be used when doing 2 reps with 3 reps in reserve
# using ballistic table
get_max_perc_1RM(
  max_reps = 3,
  RIR = 2,
  type = "ballistic"
)
# Get predicted 1RM when lifting 100kg for 5 reps to failure
get_predicted_1RM(
  weight = 100,
  reps = 5
)

# Get predicted 1RM when lifting 120kg for 3 reps with 2 reps in reserve
get_predicted_1RM(
  weight = 120,
  reps = 3,
  RIR = 2
)

# Get predicted 1RM when lifting 120kg for 2 reps with 1 reps in reserve
# using ballistic table
get_predicted_1RM(
  weight = 120,
  reps = 2,
  RIR = 1,
  type = "ballistic"
)
}
