% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_progression_table}
\alias{plot_progression_table}
\title{Plotting of th Progression Table}
\usage{
plot_progression_table(progression_table = RIR_increment, type = "\%1RM")
}
\arguments{
\item{progression_table}{Function for creating progression table. Default is \code{\link{RIR_increment}}}

\item{type}{Character string. Options include "\%1RM" (default) and "Adjustment"}
}
\value{
\code{ggplot2} object
}
\description{
Functions for creating \code{ggplot2} plot of the Progression Table
}
\examples{
plot_progression_table(RIR_increment, "\%1RM")
plot_progression_table(RIR_increment, "Adjustment")
}
