% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vertical.R
\name{plot_vertical}
\alias{plot_vertical}
\title{Plotting of the Vertical Planning}
\usage{
plot_vertical(
  vertical_plan,
  reps = c(5, 5, 5),
  font_size = 8,
  label_size = 2.5,
  ...
)
}
\arguments{
\item{vertical_plan}{Vertical Plan functions}

\item{reps}{Numeric vector}

\item{font_size}{Numeric. Default is 8}

\item{label_size}{Numeric. Default is 2.5}

\item{...}{Forwarded to \code{vertical_plan} function}
}
\description{
Function for creating \code{ggplot2} plot of the Vertical Planning function
}
\examples{
plot_vertical(vertical_block_undulating)
}
