% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression-tables-percRM.R,
%   R/progression-tables.R, R/progressiong_tables-relInt.R
\name{percMR_step_var}
\alias{percMR_step_var}
\alias{percMR_step_const}
\alias{progression_table}
\alias{RIR_increment}
\alias{RIR_increment_fixed_2}
\alias{RIR_increment_fixed_4}
\alias{perc_drop}
\alias{perc_drop_fixed_5}
\alias{perc_drop_fixed_25}
\alias{generate_progression_table}
\alias{relInt}
\title{Progression table}
\usage{
percMR_step_var(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM_percMR
)

percMR_step_const(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM_percMR
)

RIR_increment(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

RIR_increment_fixed_2(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

RIR_increment_fixed_4(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

perc_drop(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

perc_drop_fixed_5(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

perc_drop_fixed_25(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM
)

generate_progression_table(
  progression_table = RIR_increment,
  type = c("grinding", "ballistic"),
  volume = c("intensive", "normal", "extensive"),
  reps = 1:12,
  step = seq(-3, 0, 1),
  ...
)

relInt(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  adjustment = 0,
  func_max_perc_1RM = get_max_perc_1RM_relInt
)
}
\arguments{
\item{reps}{Number of reps}

\item{step}{Progression step. Default is 0. Use negative numbers (i.e., -1, -2)}

\item{volume}{Character string: 'intensive', 'normal' (Default), or 'extensive'}

\item{type}{Type of max rep table. Options are grinding (Default) and ballistic.}

\item{adjustment}{Additional adjustment. Default is 0.}

\item{func_max_perc_1RM}{What max rep table should be used?
Default is \code{\link{get_max_perc_1RM}}}

\item{progression_table}{Progression table function to use. Default is
\code{\link{RIR_increment}}}

\item{...}{Forwarded to \code{progression_table} for using different rep max function}
}
\value{
List with two elements: \code{adjustment} and \code{perc_1RM}
}
\description{
Progression table explained in the 'Strength Training Manual' Chapter 6.
}
\section{Functions}{
\itemize{
\item \code{percMR_step_var}: \%MR Step Variable progression table

\item \code{percMR_step_const}: \%MR Step Constant progression table

\item \code{RIR_increment}: RIR Increment progression table. This is the original
progression table from the Strength Training Manual

\item \code{RIR_increment_fixed_2}: Fixed 2 RIR Increment progression table. This variant have fixed RIR
increment across reps from phases to phases (2RIR) and 2RIR difference between extensive, normal, and
intensive schemes

\item \code{RIR_increment_fixed_4}: Fixed 4 RIR Increment progression table. This variant have fixed RIR
increment across reps from phases to phases (4RIR) and 4RIR difference between extensive, normal, and
intensive schemes

\item \code{perc_drop}: Percent Drop progression table. This is the original
progression table from the Strength Training Manual

\item \code{perc_drop_fixed_5}: 5\% Fixed Percent Drop progression table. This variant have fixed percent
drops across reps from phases to phases (5\%) and 5\% difference between extensive, normal, and
intensive schemes

\item \code{perc_drop_fixed_25}: 2.5\% Fixed Percent Drop progression table. This variant have fixed percent
drops across reps from phases to phases (2.5\%) and 2.5\% difference between extensive, normal, and
intensive schemes

\item \code{generate_progression_table}: Generates progression tables

\item \code{relInt}: Relative Intensity progression table
}}

\examples{
percMR_step_var(10, step = seq(-3, 0, 1))
percMR_step_var(10, step = seq(-3, 0, 1), volume = "extensive")
percMR_step_var(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(percMR_step_var)

# Plot progression table
plot_progression_table(percMR_step_var)
plot_progression_table(percMR_step_var, "adjustment")
percMR_step_const(10, step = seq(-3, 0, 1))
percMR_step_const(10, step = seq(-3, 0, 1), volume = "extensive")
percMR_step_const(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(percMR_step_const)

# Plot progression table
plot_progression_table(percMR_step_const)
plot_progression_table(percMR_step_const, "adjustment")
RIR_increment(10, step = seq(-3, 0, 1))
RIR_increment(10, step = seq(-3, 0, 1), volume = "extensive")
RIR_increment(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(RIR_increment)

# Plot progression table
plot_progression_table(RIR_increment)
plot_progression_table(RIR_increment, "adjustment")
RIR_increment_fixed_2(10, step = seq(-3, 0, 1))
RIR_increment_fixed_2(10, step = seq(-3, 0, 1), volume = "extensive")
RIR_increment_fixed_2(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(RIR_increment_fixed_2)

# Plot progression table
plot_progression_table(RIR_increment_fixed_2)
plot_progression_table(RIR_increment_fixed_2, "adjustment")
RIR_increment_fixed_4(10, step = seq(-3, 0, 1))
RIR_increment_fixed_4(10, step = seq(-3, 0, 1), volume = "extensive")
RIR_increment_fixed_4(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(RIR_increment_fixed_4)

# Plot progression table
plot_progression_table(RIR_increment_fixed_4)
plot_progression_table(RIR_increment_fixed_4, "adjustment")
perc_drop(10, step = seq(-3, 0, 1))
perc_drop(10, step = seq(-3, 0, 1), volume = "extensive")
perc_drop(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(perc_drop)

# Plot progression table
plot_progression_table(perc_drop)
plot_progression_table(perc_drop, "adjustment")
perc_drop_fixed_5(10, step = seq(-3, 0, 1))
perc_drop_fixed_5(10, step = seq(-3, 0, 1), volume = "extensive")
perc_drop_fixed_5(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(perc_drop_fixed_5)

# Plot progression table
plot_progression_table(perc_drop_fixed_5)
plot_progression_table(perc_drop_fixed_5, "adjustment")
perc_drop_fixed_25(10, step = seq(-3, 0, 1))
perc_drop_fixed_25(10, step = seq(-3, 0, 1), volume = "extensive")
perc_drop_fixed_25(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(perc_drop_fixed_25)

# Plot progression table
plot_progression_table(perc_drop_fixed_25)
plot_progression_table(perc_drop_fixed_25, "adjustment")
generate_progression_table()

generate_progression_table(type = "grinding", volume = "normal")

generate_progression_table(
  reps = 1:5,
  step = seq(-5, 0, 1),
  type = "grinding",
  volume = "normal"
)
relInt(10, step = seq(-3, 0, 1))
relInt(10, step = seq(-3, 0, 1), volume = "extensive")
relInt(5, step = seq(-3, 0, 1), type = "ballistic")

# Generate progression table
generate_progression_table(relInt)

# Plot progression table
plot_progression_table(relInt)
plot_progression_table(relInt, "adjustment")
}
