% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progression-tables.R
\name{progression_table}
\alias{progression_table}
\alias{RIR_increment}
\alias{perc_drop}
\alias{generate_progression_table}
\title{Progression table}
\usage{
RIR_increment(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  func_max_perc_1RM = get_max_perc_1RM
)

perc_drop(
  reps,
  step = 0,
  volume = "normal",
  type = "grinding",
  func_max_perc_1RM = get_max_perc_1RM
)

generate_progression_table(
  type = c("grinding", "ballistic"),
  volume = c("intensive", "normal", "extensive"),
  reps = 1:12,
  step = seq(-3, 0, 1),
  func_max_perc_1RM = get_max_perc_1RM
)
}
\arguments{
\item{reps}{Number of reps}

\item{step}{Progression step. Default is 0. Use negative numbers (i.e., -1, -2)}

\item{volume}{Character string: 'intensive', 'normal' (Default), or 'extensive'}

\item{type}{Type of max rep table. Options are grinding (Default) and ballistic.}

\item{func_max_perc_1RM}{What max rep table should be used?
Default is \code{\link{get_max_perc_1RM}}}
}
\value{
List with two elements: \code{adjustment} and \code{perc_1RM}
}
\description{
Progression table explained in the 'Strength Training Manual' Chapter 6.
}
\section{Functions}{
\itemize{
\item \code{RIR_increment}: RIR Increment progression table

\item \code{perc_drop}: Percent Drop progression table

\item \code{generate_progression_table}: Generates RIR Inc and Perc Drop progression tables
}}

\examples{
RIR_increment(10, step = seq(-3, 0, 1))
RIR_increment(10, step = seq(-3, 0, 1), volume = "extensive")
RIR_increment(5, step = seq(-3, 0, 1), type = "ballistic")
perc_drop(10, step = seq(-3, 0, 1))
perc_drop(10, step = seq(-3, 0, 1), volume = "extensive")
perc_drop(5, step = seq(-3, 0, 1), type = "ballistic")
generate_progression_table()

generate_progression_table(type = "grinding", volume = "normal")

generate_progression_table(
  reps = 1:5,
  step = seq(-5, 0, 1),
  type = "grinding",
  volume = "normal"
)
}
